Glitch logic d3 superjump and d6 elephant bomb trigger (#127)
* d3 glitch logic: superjump to boots chest

Use the pathway that leads to the nightmare to superjump up north towards the pegasus boots chest after the miniboss. This skips the miniboss requirement for the chest and only requires you to open 2 key block out of 4 leading to the nightmare room

* d6 bomb trigger elephant statues

you can also bomb trigger the elephant statue next to the chest near the entrance that blocks your way to the second chest to break it. This also fixes hell logic after miniboss to recognize you can break the elephant statue with bombs as well which it didnt before