[hard] bottom left vire needs additional requirements for just bombs
Entering d8 with just bombs leads to a scenario where you'd need more than 10 bombs (barring really dumb stuff) to get the vire item. Deal with it by forcing bomb vire only on the route that bypasses hinox. Technically Shield + Bomb would be feasible at exact 10 but probably better to not include that incase a bomb is required to reach the dungeon.