fix advanced logic issues and recent community decisions/findings (#149)
* [hard] traverse mountain connector to d7 with just rooster

Grab the rooster while on top of the pit to avoid having to use the hookshot to cross the gap

* [hell] unclipped shaq jumps against corner blocks

apparently the corner blocks can be shaq jumped against, but not consistently (similar to naked superjumps I guess). Seems to work similar to naked superjumps with sword out, so fine to add it to hell logic.

* [hell] hookshot tricks to pass through blocks

skip keyblock with some well tricked hookshots, and get to right side of crossroads without feather or boots in niche scenario's by clipping through the blocks

* [hard] bottom left vire needs additional requirements for just bombs

Entering d8 with just bombs leads to a scenario where you'd need more than 10 bombs (barring really dumb stuff) to get the vire item. Deal with it by forcing bomb vire only on the route that bypasses hinox. Technically Shield + Bomb would be feasible at exact 10 but probably better to not include that incase a bomb is required to reach the dungeon.

* [hell] swordless log cave crystal chest

use a superjump + hookshot clipping with 3 keese to gain enough height to sideways block push the block down to gain access to chest

* swordless log cave crystal chest can be done with a naked superjump

does not need bow/magic rod

* [hell] animal village connector pass without boots

Unclipped shaq jump from either top wall or bottom wall depending on direction, and als othe option to hookshot clip through using the bottom fish if coming from the left side