Add a script to launch the LADXR web UI locally (issue #56)
The script sets up a simple local server with standard Python tools and
serves the web UI. It then launches a browser window with the main page.

Note that this relies on the basic directory structure of the served
files, and the URLS under which they are requested, to remain the
same:

* The contents of the www folder being served at top level
* The contents of the gfx folder being served at LADXR/gfx/
* The tgz file being served at /js/ladxr.tar.gz

So any restructuring will require adjustments in the launcher
script. Changes soley within the www and gfx folders are no problem.