Cleaned up text in user interface
Mostly just punctuation and formatting fixes, along with cleaning up some typos. Also documents options that were overlooked. As an aside, probably should make the text box wider for at least the Goal setting, as there is a lot of text for a lot of options and it no longer fits on the page...