Cleaned up text in user interface and add new graphic set (#180)
* Cleaned up text in user interface

Mostly just punctuation and formatting fixes, along with cleaning up some typos. Also documents options that were overlooked. As an aside, probably should make the text box wider for at least the Goal setting, as there is a lot of text for a lot of options and it no longer fits on the page...

* Added Saria

New graphics set! Note that this doesn't update the UI.

* Dang it, fine, going all in then

Sorry Daid, I wanted to keep this separate, but Git don't work that way

* Update settings.py

* WHOOPS

* Update settings.py